Update README.md
This project defines an array data model for a hotel search tool that allows you to filter the type of hotel you want to stay in. The Filter component has a checkbox, a name, and a number representing the number of hotels matching the filter. It also has functionality for exporting, importing, and prop passing to build the UI.